requirements.txt
numpy>=1.21.0
scikit-learn>=1.0.0
tensorflow>=2.8.0
pyyaml>=6.0
flask>=2.0.0
requests>=2.27.0
pandas>=1.3.0
matplotlib>=3.5.0
